bruh

+++ gestione/amministrazione/dipendenti/model_dipendenti/model_dipendenti.py
@@ -4,12 +4,11 @@ import pickle
 
 class Insieme_Dipendenti():
 
+
     def __init__(self):
         self.lista_dipendenti = []
-        if os.path.isfile("gestione/amministrazione/dipendenti/data_dipendenti/lista_dipendenti_salvata.pickle"):
-
-            with open("gestione/amministrazione/dipendenti/data_dipendenti/lista_dipendenti_salvata.pickle", "rb") as file:
-
+        if os.path.isfile("amministrazione/dipendenti/data_dipendenti/lista_dipendenti_salvata.pickle"):
+            with open("amministrazione/dipendenti/data_dipendenti/lista_dipendenti_salvata.pickle", "rb") as file:
                 self.lista_dipendenti = pickle.load(file)
 
     def aggiungi_dipendente(self, dipendente):
@@ -28,13 +27,12 @@ class Insieme_Dipendenti():
         return self.lista_dipendenti
 
     def get_dipendente_by_id(self, id):
-
         for dipendente in self.lista_dipendenti:
             if dipendente.id == id:
                 return dipendente
         return None
 
     def save_data(self):
-
-        with open("gestione/amministrazione/dipendenti/data_dipendenti/lista_dipendenti_salvata.pickle", "wb") as handle:
-            pickle.dump(self.lista_dipendenti, handle, pickle.HIGHEST_PROTOCOL)
+        if os.path.isfile("amministrazione/dipendenti/data_dipendenti/lista_dipendenti_salvata.pickle"):
+            with open("amministrazione/dipendenti/data_dipendenti/lista_dipendenti_salvata.pickle", "wb") as handle:
+                pickle.dump(self.lista_dipendenti, handle, pickle.HIGHEST_PROTOCOL)
